[CHG] Shop and Item management

+++ ProductTracker/ProductTracker.Web/ManageItems.aspx.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Collections.Generic;
 using System.Data;
+using System.Web;
 
 #endregion
 
@@ -36,16 +37,34 @@ namespace ProductTracker.Web
                     LabelItemsSelect.Text = settingsManager.GetSettingValue("LabelItemsSelect", setting);
                     LinkButtonItemsSelect.Text = settingsManager.GetSettingValue("LinkButtonItemsSelect", setting);
                     LinkButtonItemsAddNew.Text = settingsManager.GetSettingValue("LinkButtonItemsAddNew", setting);
-                    LabelItemId.Text = settingsManager.GetSettingValue("LabelItemId", setting);
                     LabelItemName.Text = settingsManager.GetSettingValue("LabelItemName", setting);
                     LabelItemNote.Text = settingsManager.GetSettingValue("LabelItemNote", setting);
                     LabelItemItemType.Text = settingsManager.GetSettingValue("LabelItemItemType", setting);
                     LinkButtonMiddleBodySubmit.Text = settingsManager.GetSettingValue("LinkButtonMiddleBodySubmit",
                                                                                       setting);
+                    LinkButtonMiddleBodyUpdate.Text = settingsManager.GetSettingValue("LinkButtonMiddleBodyUpdate",
+                                                                                      setting);
                     LinkButtonMiddleItemTypesDelete.Text =
                         settingsManager.GetSettingValue("LinkButtonMiddleItemTypesDelete", setting);
                 }
             }
+            EnableNew();
+        }
+
+        private void EnableNew()
+        {
+            LinkButtonMiddleBodySubmit.Visible = true;
+            LinkButtonMiddleBodyUpdate.Visible = false;
+            LinkButtonMiddleItemTypesDelete.Visible = false;
+            //DropDownListItemsSelect.Enabled = true;
+        }
+
+        private void EnableUpdate()
+        {
+            LinkButtonMiddleBodySubmit.Visible = false;
+            LinkButtonMiddleBodyUpdate.Visible = true;
+            LinkButtonMiddleItemTypesDelete.Visible = true;
+            //DropDownListItemsSelect.Enabled = false;
         }
 
         private void PopulateItems()
@@ -68,13 +87,25 @@ namespace ProductTracker.Web
         {
             string selectedValue = DropDownListItemsSelect.SelectedValue;
             DataBase dataBase = new DataBase();
-            Item item = dataBase.GetItem(new Guid(selectedValue));
+            Item item = dataBase.GetItem(selectedValue);
+            if (item == null)
+            {
+                return;
+            }
             PopulateItemTypes();
             TextBoxMiddleBodyId.Text = item.Id;
             TextBoxMiddleBodyName.Text = item.Name;
             TextBoxMiddleBodyNotes.Text = item.Notes;
             DropDownListMiddleBodyItemType.SelectedValue = item.ItemType.ToString();
-            TextBoxMiddleBodyUniqueId.Text = item.UniqueId.ToString();
+            EnableUpdate();
+            HttpCookie cookie = Request.Cookies[CookieItemuniqueid];
+            if (cookie == null)
+            {
+                cookie = new HttpCookie(CookieItemuniqueid);
+            }
+            cookie["Name"] = item.UniqueId.ToString();
+            cookie.Expires = DateTime.Now.AddMinutes(15);
+            Response.Cookies.Add(cookie);
         }
 
         protected void LinkButtonMiddleBodySubmit_Click(object sender, EventArgs e)
@@ -85,15 +116,7 @@ namespace ProductTracker.Web
             string notes = TextBoxMiddleBodyNotes.Text;
             string itemType = DropDownListMiddleBodyItemType.SelectedValue;
             Item item = new Item(id, name) {Notes = notes, ItemType = Int32.Parse(itemType)};
-            if (TextBoxMiddleBodyUniqueId.Text.Length > 8)
-            {
-                item.UniqueId = new Guid(TextBoxMiddleBodyUniqueId.Text.Trim());
-                dataBase.UpdateItem(item);
-            }
-            else
-            {
-                dataBase.InsertItem(item);
-            }
+            dataBase.InsertItem(item);
             PopulateItems();
         }
 
@@ -110,18 +133,54 @@ namespace ProductTracker.Web
 
         protected void LinkButtonMiddleItemTypesDelete_Click(object sender, EventArgs e)
         {
+            HttpCookie cookie = Request.Cookies[CookieItemuniqueid];
+            if (cookie == null)
+            {
+                return;
+            }
+            string id = cookie["Name"];
             DataBase dataBase = new DataBase();
-            string id = TextBoxMiddleBodyUniqueId.Text;
             dataBase.DeleteItem(id);
             PopulateItems();
         }
 
         protected void LinkButtonItemsAddNew_Click(object sender, EventArgs e)
         {
+            ClearFields();
+        }
+
+        private void ClearFields()
+        {
             TextBoxMiddleBodyId.Text = "";
             TextBoxMiddleBodyName.Text = "";
             TextBoxMiddleBodyNotes.Text = "";
-            TextBoxMiddleBodyUniqueId.Text = "";
+            EnableNew();
         }
+
+        protected void LinkButtonMiddleBodyUpdate_Click(object sender, EventArgs e)
+        {
+            HttpCookie cookie = Request.Cookies[CookieItemuniqueid];
+            if (cookie == null)
+            {
+                return;
+            }
+            DataBase dataBase = new DataBase();
+            string id = TextBoxMiddleBodyId.Text;
+            string name = TextBoxMiddleBodyName.Text;
+            string notes = TextBoxMiddleBodyNotes.Text;
+            string itemType = DropDownListMiddleBodyItemType.SelectedValue;
+            Item item = new Item(id, name)
+                {
+                    Notes = notes, 
+                    ItemType = Int32.Parse(itemType), 
+                    UniqueId = new Guid(cookie["Name"]),
+                };
+            dataBase.UpdateItem(item);
+            ClearFields();
+            PopulateItems();
+
+        }
+
+        private const string CookieItemuniqueid = "ItemUniqueId";
     }
 }
